Giorno 18/esempi_django/requirements.txt
django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
openai==1.66.1
python-dotenv==1.0.1
